# public/openapi.yaml
openapi: 3.1.0
info:
  title: Meditherapy Skin Coach Plugin API
  description: |
    메디테라피 스킨 코치 API. 피부고민 기반 루틴 추천, 익명 체크인 저장, Convex 기반 사용 데이터 요약을 제공합니다.
    개인정보는 저장하지 않고 구조화된 익명 이벤트만 저장합니다.
  version: 1.1.0
servers:
  - url: https://meditherapy-skin-coach-dashboard.vercel.app
paths:
...
  /api/record-event:
    post:
      operationId: recordSkinCoachEvent
      summary: 익명 루틴/체크인 이벤트 저장
      description: 이름, 전화번호, 이메일, 주소, 주문번호, 원문 대화, 사진 원본을 받지 않고 익명 구조화 이벤트만 Convex DB에 저장합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                eventType:
                  type: string
                  enum: [routine_recommendation, task_schedule_created, checkin, routine_adjustment, product_feedback, ontology_gap]
                userSeed:
                  type: string
                  description: 익명 해시 생성을 위한 비식별 seed. 원문 개인정보 사용 금지.
                skinType:
                  type: string
                concerns:
                  type: array
                  items:
                    type: string
                sensitivity:
                  type: string
                routineId:
                  type: string
                productIds:
                  type: array
                  items:
                    type: string
                checkinDay:
                  type: number
                drynessScore:
                  type: number
                irritationScore:
                  type: number
                satisfactionScore:
                  type: number
                action:
                  type: string
                locale:
                  type: string
                ontologyGapFields:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: 저장 결과
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  id:
                    type: string
                  saved:
                    type: object
...